+++ jobs/generic/sql_pandas_job.py
@@ -1,5 +1,5 @@
 from yaetos.etl_utils import Commandliner
-from yaetos.sql_job_pandas import Job
+from yaetos.sql_pandas_job import Job
 
 
 if __name__ == "__main__":
