+++ jobs/generic/sql_spark_job.py
@@ -1,5 +1,5 @@
 from yaetos.etl_utils import Commandliner
-from yaetos.sql_job import Job
+from yaetos.sql_spark_job import Job
 
 
 if __name__ == "__main__":
